+++ lib/ts/recipe/thirdparty/components/features/signInAndUpCallback/index.tsx
@@ -29,6 +29,7 @@ import {} from "../../../types";
 import { SignInAndUpCallbackTheme } from "../../themes/signInAndUpCallback";
 import Recipe from "../../../recipe";
 import { ComponentOverrideContext } from "../../../../../components/componentOverride/componentOverrideContext";
+import { defaultTranslationsThirdParty } from "../../themes/translations";
 
 type PropType = FeatureBaseProps & { recipe: Recipe };
 
@@ -106,7 +107,10 @@ class SignInAndUpCallback extends PureComponent<PropType, unknown> {
 
         return (
             <ComponentOverrideContext.Provider value={componentOverrides}>
-                <FeatureWrapper useShadowDom={this.props.recipe.config.useShadowDom} isEmbedded={this.getIsEmbedded()}>
+                <FeatureWrapper
+                    useShadowDom={this.props.recipe.config.useShadowDom}
+                    isEmbedded={this.getIsEmbedded()}
+                    defaultStore={defaultTranslationsThirdParty}>
                     <StyleProvider
                         rawPalette={this.props.recipe.config.palette}
                         defaultPalette={defaultPalette}